fix_bug

+++ HotelBooking/database/migrations/2022_07_19_034616_create_rooms_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('rooms', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->float('price');
+            $table->integer('price');
             $table->text('description');
             $table->string('convenient');
             $table->string('image_path');
